renamed search_result_actions to youtube_video_actions
